backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-jose[cryptography]>=3.3.0
supabase>=2.3.0
openai>=1.10.0
anthropic>=0.18.0
google-generativeai>=0.4.0
pydantic>=2.5.0
python-dotenv>=1.0.0
httpx>=0.26.0
